add searchRequest.ts
