+++ AGENTS.md
@@ -97,6 +97,7 @@ app/frontend/
 - アプリ状態は `app/frontend/src/store/StoreManager.ts` を正として扱う。
 - 検索条件のURL反映や初期復元は `app/frontend/src/store/searchManager.ts` と `initializeApp.ts` を確認する。
 - API通信は `app/frontend/src/api/fetchData.ts` など既存API層に合わせる。
+- 検索リクエストのURL・HTTPオプション生成は `app/frontend/src/api/searchRequest.ts` に置き、`fetchData.ts` は取得実行、レスポンス処理、Store反映に集中させる。
 - StoreはAPIを直接呼ばない。仮想スクロールで未取得ページが必要な場合も、コンポーネントから `searchManager.requestNextPage()` を呼び、fetch の起動は `searchManager` / `api/fetchData.ts` 側へ集約する。
 - Simple Search条件のdefault差分抽出は `store/simpleSearchConditions.ts` に置く。`api/fetchData.ts` から `searchManager.ts` を import すると循環依存になるため避ける。
 - コンポーネントやViewから直接URLやfetchの仕様を増やす前に、既存のStore/API層へ寄せられるか確認する。
